MachineLearningStudio/requirements.txt
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3
scikit-learn==1.6.1
pillow==11.2.1
joblib==1.5.0